+++ solutions/bmi_calculator.py
