+++ solutions/water_intake_solution.py
